+++ scripts/installer/buildInstaller.sh
@@ -1,3 +1,0 @@
-cd ../../
-docker run --rm -i -v $PWD:/work -w /work/scripts/installer amake/innosetup installer.iss
-cd scripts/installer