+++ scripts/windows-setup/build.sh
@@ -1,0 +1,3 @@
+cd ../../
+docker run --rm -i -v $PWD:/work -w /work/scripts/windows-setup amake/innosetup installer.iss
+cd scripts/windows-setup