+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Packs on CurseForge using JumpLoader no longer use Forge [#444]
 - Add OS information to Google Analytics
 - Use newer universalJavaApplicationStub for OSX app
+- Add in debian packaging
 
 ### Fixes
 - Dialog when installing file from CurseForge not showing progress